Restore previous virtio_pnp_match structure
I was under the impression an extra field would need to be added to this structure, used for building "device-identifying data" via MODULE_PNP_INFO, to distinguish virtio_blk and virtio_blk_iocap. This is not the case, and I don't want to rock the boat with unnecessary changes.
This also means the standard VIRTIO_SIMPLE_PNPINFO macro can be used in virtio_blk_iocap. Hooray!